Remove Meta from Index HTML
Closes #1

Enabling the "Beta" Pre-rendering feature in Netlify seems to solve rendering `<meta>` tags. If a bot user-agent is detected (eg. Twitterbot), Netlify will use a headless browser to render the page and then serve the HTML to the bot. 

SSR may be needed in the future, but for now, the issue is resolved.